+++ src/disassembler/operands.h
@@ -9,6 +9,6 @@ unsigned char getSizeOfOperand(struct Operand* operand);
 
 unsigned char compareOperands(struct Operand* op1, struct Operand* op2);
 
-unsigned char isOperandStackVar(struct Operand* operand, int stackFrameSize);
+unsigned char isOperandStackVar(struct Operand* operand, long long stackFrameSize);
 
-unsigned char isOperandStackArg(struct Operand* operand, int stackFrameSize);
+unsigned char isOperandStackArg(struct Operand* operand, long long stackFrameSize);
